requirements.txt
paddleocr>=2.7.0
paddlepaddle
pymupdf
pillow
img2pdf
transformers
torch
sentencepiece
argostranslate
